+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0] - 2026-09-12
+### Added
+- New view "seasonal" to show the seasonal fruits and vegetables in the home
+  location and in all cities of interest, so now the user could advise their
+  friends and relatives what seasonal fruits/veggies they should eat for a
+  healthy lifestyle
+
+### Fixed
+- The seasonal fruits/veggies information now excludes the major non-agricultural
+  zones, like Antarctica, the Sahara desert etc.
+
 ## [1.4.2] - 2026-09-11
 ### Fixed
 - Added measurement units to FIREBALLS warning. Added a safeguard check for future API version.